Add files via upload
I tried to correct a few of the lines of code and increase the number of iterations for the split step method, and now it seems that even with a contact potential and dipole dipole interactions the solution converges to the original ground state (no interactions) with enough iterations. But if the number of iterations is lower, the effects of the interactions are present.